Fix ex15

+++ pl05b/ex15/main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "soma.h"
 
-int temp=0, temp_p = 0, res = 0;
+short temp=0, temp_p = 0, res = 0;
 
 int main(void){
 	printf("Temperatura actual:");
